Backend/agents/eror_scheduling/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
xgboost==2.0.3
ortools==9.8.3296
mlflow==2.9.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
joblib==1.3.2

